Adding first cut to add diagrams
